README
This app will show matches for matchday 6 with an endpoint: "https://api.football-data.org/v2/competitions/PL/matches?matchday=6". I have taken the API Key out of my code to secure the data.

Open the project using Xcode, enter in the unique API Key where every URL has been inputted. (There should only be 3 request.setValue functions where you can insert the API key). Run the project on any simulator.

I chose to start by requesting the data from the URL link with the Authorization token. This made it easy to see the layout of the JSON file and allow me to structure my dictionary. Once the data was parsing into the console, I started picking out the names of home teams and away teams. I used JSONSerialization which converts between JSON and objects.

ViewController
This is where the main screen code is held, i have made a function call from viewDidLoad to parseData which gets the data and stores it into a local variable array. I used a table view with prototype cell to organise the data being published to the app. This is connected by a IBOutlet called myTableView. The labels were connected to a cocoa touch file named TableViewCell.

MatchViewController
Once you click on the match you want to choose from the Table View, it will send you to another view controller. This controller hold images and the team names of home and away. I had some difficulty adding more information as the url didn't seem to work when i loaded https://api.football-data.org/v2/competitions/PL/teams. I also managed to get the data which was stored in the array into the second view controller.

TableViewController
I tried to access the league table for the premier league but the data didn't return back whenever i tried to dump it to console. I decided to leave this in the app to show you where i got to and if there are any ideas on what i could fix.